单元测试

+++ myopia-business/business-api/screening-app-api/src/main/java/com/wupol/myopia/business/api/screening/app/domain/dto/BloodPressureDTO.java
@@ -24,12 +24,12 @@ public class BloodPressureDTO  extends ScreeningResultBasicData {
     /**
      * 舒张压
      */
-    private Float dbp;
+    private Integer dbp;
 
     /**
      * 收缩压
      */
-    private Float sbp;
+    private Integer sbp;
 
     @Override
     public VisionScreeningResult buildScreeningResultData(VisionScreeningResult visionScreeningResult) {
